Revert python -> sympy_code (closes gh-12807)
